# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Yang's Fried Rice - System API",
    "version": "1.0.0",
    "description": "Interactive Swagger/OpenAPI documentation for Menu Management, Administration, Order Oversight, and Reporting. Use 'Try it out' to execute live requests."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/api/admin/employees": {
      "get": {
        "tags": [
          "Admin - Employees"
        ],
        "summary": "Get all employees",
        "description": "Lists all registered employee accounts. Requires admin or manager role.",
        "operationId": "getEmployees",
        "responses": {
          "200": {
            "description": "List of employees",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer",
                      "example": 3
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Employee"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Admin - Employees"
        ],
        "summary": "Create employee",
        "description": "Creates a Supabase Auth user and corresponding employee record. Requires admin or manager role.",
        "operationId": "createEmployee",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEmployeeInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Employee created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Employee created successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Employee"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/employees/{id}": {
      "delete": {
        "tags": [
          "Admin - Employees"
        ],
        "summary": "Delete employee by ID",
        "description": "Deletes the employee record and Auth user. Prevents self-deletion. Requires admin role.",
        "operationId": "deleteEmployee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Employee ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Employee deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Employee deleted successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "employee_id": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/employees/{id}/role": {
      "patch": {
        "tags": [
          "Admin - Employees"
        ],
        "summary": "Change employee role",
        "description": "Updates employee role with role hierarchy enforcement (admin > manager > staff > rider). Requires admin or manager.",
        "operationId": "changeEmployeeRole",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Employee ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRoleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Role updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Employee role updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Employee"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/customers": {
      "get": {
        "tags": [
          "Admin - Customers"
        ],
        "summary": "Get all customers",
        "description": "Lists all registered customer accounts including active and disabled status. Requires admin or manager.",
        "operationId": "getCustomers",
        "responses": {
          "200": {
            "description": "List of customers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "integer",
                      "example": 10
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Customer"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/customers/{id}": {
      "patch": {
        "tags": [
          "Admin - Customers"
        ],
        "summary": "Update customer or toggle disabled",
        "description": "Updates customer details or enables/disables the account via is_account_disabled. Requires admin or manager.",
        "operationId": "updateCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Customer ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customer updated successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Customer"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Admin - Customers"
        ],
        "summary": "Delete customer by ID",
        "description": "Deletes customer record and associated Auth user. Requires admin role.",
        "operationId": "deleteCustomer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Customer ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Customer deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Customer deleted successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "customer_id": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/employees/{id}/disable": {
      "patch": {
        "tags": [
          "Admin - Employees"
        ],
        "summary": "Disable or enable employee account",
        "description": "Temporarily disables or re-enables an employee account. Enforces role hierarchy: Admin can disable any employee (except self); Manager can disable Staff and Rider; Staff/Rider cannot disable anyone.",
        "operationId": "toggleEmployeeDisabled",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Employee ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "is_account_disabled"
                ],
                "properties": {
                  "is_account_disabled": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee disabled/enabled successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Employee account disabled successfully"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Employee"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden by hierarchy or self-disable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/employees/{id}/password": {
      "patch": {
        "tags": [
          "Admin - Employees"
        ],
        "summary": "Reset employee password",
        "description": "Allows higher-ups to reset another employee's password with role hierarchy enforcement: ADMIN can reset any employee; MANAGER can reset STAFF and RIDER; STAFF/RIDER cannot reset others.",
        "operationId": "resetEmployeePassword",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Employee ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Employee password reset successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Employee password updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "employee_id": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden by role hierarchy",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Employee not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "Internal server error"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
...
      "Employee": {
        "type": "object",
        "required": [
          "employee_id",
          "name",
          "email",
          "role"
        ],
        "properties": {
          "employee_id": {
            "type": "string",
            "format": "uuid",
            "example": "e0000000-0000-0000-0000-000000000001"
          },
          "name": {
            "type": "string",
            "example": "Juan Dela Cruz"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "juan@yangsfriedrice.com"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "STAFF",
              "RIDER"
            ],
            "example": "ADMIN"
          },
          "last_access_log": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-09-10T08:30:00Z"
          },
          "is_account_disabled": {
            "type": "boolean",
            "example": false
          }
        }
      },
      "CreateEmployeeInput": {
        "type": "object",
        "required": [
          "name",
          "email",
          "password",
          "role"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Maria Santos"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "maria@yangsfriedrice.com"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "example": "SecurePass123!"
          },
          "role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "STAFF",
              "RIDER"
            ],
            "example": "ADMIN"
          }
        }
      },
      "ChangeRoleInput": {
        "type": "object",
        "required": [
          "new_role"
        ],
        "properties": {
          "new_role": {
            "type": "string",
            "enum": [
              "ADMIN",
              "MANAGER",
              "STAFF",
              "RIDER"
            ],
            "example": "MANAGER"
          }
        }
      },
      "Customer": {
        "type": "object",
        "required": [
          "customer_id",
          "name",
          "email"
        ],
        "properties": {
          "customer_id": {
            "type": "string",
            "format": "uuid",
            "example": "c0000000-0000-0000-0000-000000000001"
          },
          "name": {
            "type": "string",
            "example": "Maria Clara"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "maria.clara@example.com"
          },
          "phone_number": {
            "type": "string",
            "nullable": true,
            "example": "+639171234567"
          },
          "is_account_disabled": {
            "type": "boolean",
            "example": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2026-09-01T12:00:00Z"
          }
        }
      },
      "UpdateCustomerInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100,
            "example": "Maria Clara Updated"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "maria.updated@example.com"
          },
          "phone_number": {
            "type": "string",
            "nullable": true,
            "example": "+639171234567"
          },
          "is_account_disabled": {
            "type": "boolean",
            "example": false
          }
        }
      },
...
      "ChangePasswordInput": {
        "type": "object",
        "required": [
          "new_password"
        ],
        "properties": {
          "new_password": {
            "type": "string",
            "minLength": 8,
            "example": "NewSecurePass123!"
          }
        }
      }
    }
  }
}